implementing the DAO for nodeJS
